refactor: split third app.rs submodules under ADR-028
