+++ src/app.rs
@@ -50,11 +50,14 @@ use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
 
 mod commands;
+mod inline;
 mod overlay;
 mod scroll;
+mod shell;
 #[cfg(test)]
 mod tests;
 mod turn;
+mod turn_start;
 pub(crate) mod util;
 
 use self::overlay::summarize_tool_approval_context;
@@ -771,274 +774,6 @@ impl TuiMode {
         Self::registered_slash_command(input)
             .map(|(spec, _)| matches!(spec.id, SlashCommandId::Edit | SlashCommandId::Fix))
             .unwrap_or(false)
-    }
-
-    fn expand_inline_file_tokens(&self, input: &str) -> String {
-        if input.starts_with('/') {
-            return input.to_string();
-        }
-
-        let assembler = ContextAssembler::default();
-        let operator = ToolOperator::new(self.working_dir.clone());
-        let mut output = String::new();
-        let mut token = String::new();
-
-        for ch in input.chars() {
-            if ch.is_whitespace() {
-                if !token.is_empty() {
-                    output.push_str(&self.expand_inline_token(&token, &operator, &assembler));
-                    token.clear();
-                }
-                output.push(ch);
-            } else {
-                token.push(ch);
-            }
-        }
-
-        if !token.is_empty() {
-            output.push_str(&self.expand_inline_token(&token, &operator, &assembler));
-        }
-
-        output
-    }
-
-    fn expand_inline_token(
-        &self,
-        token: &str,
-        operator: &ToolOperator,
-        assembler: &ContextAssembler,
-    ) -> String {
-        let Some(path) = token.strip_prefix('@') else {
-            return token.to_string();
-        };
-
-        if path.is_empty() {
-            return token.to_string();
-        }
-
-        match operator.existing_path(path) {
-            Ok(Some(resolved)) if resolved.is_dir() => {
-                match operator.list_files(Some(path), assembler.max_related) {
-                    Ok(listing) => format_inline_block("dir", path, &listing, false, None),
-                    Err(error) => format!("[dir: {path} \u{2014} {error}]"),
-                }
-            }
-            Ok(Some(_)) => match operator.read_file(path) {
-                Ok(content) => {
-                    let (content, truncated) =
-                        truncate_head_bytes(&content, assembler.max_file_bytes);
-                    format_inline_block(
-                        "file",
-                        path,
-                        &content,
-                        truncated,
-                        Some(assembler.max_file_bytes),
-                    )
-                }
-                Err(error) => format!("[file: {path} \u{2014} {error}]"),
-            },
-            Ok(None) => format!("[file: {path} \u{2014} not found]"),
-            Err(error) => format!("[file: {path} \u{2014} {error}]"),
-        }
-    }
-
-    fn handle_bang_command(&mut self, command: &str, ctx: &RuntimeContext) {
-        let command = command.trim();
-        if command.is_empty() {
-            self.push_history_line("[shell] usage: !<command>".to_string());
-            return;
-        }
-
-        if self.overlay_state.auto_approve_session {
-            self.push_history_line("[auto-approved tool: run_command session]".to_string());
-            self.start_command_session(command.to_string(), ctx);
-            return;
-        }
-
-        if let Some(scope) = self
-            .current_task
-            .active_grants
-            .get(&Capability::RunCommand)
-            .copied()
-        {
-            if matches!(scope, ApprovalScope::Once) {
-                self.current_task
-                    .active_grants
-                    .remove(&Capability::RunCommand);
-            }
-            self.push_history_line(format!(
-                "[auto-approved tool: run_command {} grant]",
-                scope_to_label(scope)
-            ));
-            self.start_command_session(command.to_string(), ctx);
-            return;
-        }
-
-        let summary = summarize_tool_approval_context("run_command", command);
-        self.push_history_line(format!("[tool approval requested: {summary}]"));
-        self.overlay_state.pending_approval = Some(PendingApproval {
-            tool_name: "run_command".to_string(),
-            input_preview: command.to_string(),
-            action: PendingApprovalAction::InlineCommand(PendingInlineCommand {
-                command: command.to_string(),
-            }),
-        });
-    }
-
-    fn start_command_session(&mut self, command: String, ctx: &RuntimeContext) {
-        let starting_batch = self.command_sessions.is_empty();
-        self.history_state.turn_in_progress = true;
-        self.history_state.cancel_pending = false;
-        self.history_state.active_assistant_index = None;
-        if starting_batch {
-            self.begin_turn_capture(format!("!{command}"));
-        }
-        let session_id = self.begin_command_session(command.clone());
-
-        let ctx = ctx.clone();
-        let cancel = ctx.turn_cancellation_token();
-        let working_dir = self.working_dir.clone();
-        tokio::spawn(async move {
-            let runner = DefaultCommandRunner::new();
-            // User-initiated !command execution — the Capability::RunCommand
-            // approval gate is the security boundary.  PassthroughSandbox is
-            // intentional here; a future ADR-024 follow-up may thread the
-            // operator-configured sandbox driver through TuiMode.
-            let request = match PassthroughSandbox
-                .wrap(shell_command_request(command.clone(), working_dir))
-            {
-                Ok(request) => request,
-                Err(error) => {
-                    ctx.emit_transcript_line(format!("[command session] error: {error}"));
-                    ctx.emit_command_session_finished(session_id);
-                    ctx.emit_turn_complete();
-                    return;
-                }
-            };
-            let (output_tx, mut output_rx) = mpsc::channel(128);
-            let mut handle = match runner.run_streaming(request, output_tx).await {
-                Ok(handle) => handle,
-                Err(error) => {
-                    ctx.emit_transcript_line(format!("[command session] error: {error}"));
-                    ctx.emit_command_session_finished(session_id);
-                    ctx.emit_turn_complete();
-                    return;
-                }
-            };
-            ctx.emit_command_session_attached(session_id, handle.pid());
-            ctx.emit_transcript_line(format_command_session_started(&command, handle.pid()));
-
-            let mut cancel_requested = false;
-            loop {
-                tokio::select! {
-                    _ = cancel.cancelled(), if !cancel_requested => {
-                        cancel_requested = true;
-                        let _ = handle.cancel();
-                        ctx.emit_transcript_line("[command session cancellation requested]".to_string());
-                    }
-                    chunk = output_rx.recv() => {
-                        match chunk {
-                            Some(chunk) => {
-                                for line in format_command_session_output(chunk) {
-                                    ctx.emit_transcript_line(line);
-                                }
-                            }
-                            None => break,
-                        }
-                    }
-                }
-            }
-
-            match handle.wait().await {
-                Ok(result) => {
-                    if cancel_requested {
-                        ctx.emit_transcript_line(format_command_session_cancelled());
-                    } else {
-                        ctx.emit_transcript_line(format_command_session_exit(result.exit_code));
-                    }
-                }
-                Err(error) => {
-                    ctx.emit_transcript_line(format!("[command session] error: {error}"));
-                }
-            }
-            ctx.emit_command_session_finished(session_id);
-            ctx.emit_turn_complete();
-        });
-    }
-
-    fn start_single_turn(
-        &mut self,
-        rendered: String,
-        ctx: &mut RuntimeContext,
-        read_only: bool,
-        supplementary_system_prompt: Option<&str>,
-    ) {
-        self.start_single_turn_with_policy(
-            rendered,
-            ctx,
-            read_only,
-            supplementary_system_prompt,
-            TurnToolPolicy::Default,
-        );
-    }
-
-    fn start_single_turn_with_policy(
-        &mut self,
-        rendered: String,
-        ctx: &mut RuntimeContext,
-        read_only: bool,
-        supplementary_system_prompt: Option<&str>,
-        turn_tool_policy: TurnToolPolicy,
-    ) {
-        self.history_state.active_assistant_index = Some(self.history_state.lines.len() - 1);
-        self.history_state.turn_in_progress = true;
-        self.read_only_turn_active = read_only;
-        self.begin_turn_capture(rendered.clone());
-        #[cfg(test)]
-        {
-            self.last_turn_input = Some(rendered.clone());
-        }
-        ctx.start_turn_with_system_prompt_and_policy(
-            rendered,
-            supplementary_system_prompt.map(ToString::to_string),
-            turn_tool_policy,
-        );
-    }
-
-    fn selected_system_prompt(&self) -> &'static str {
-        self.model_profile
-            .system_prompt_text()
-            .unwrap_or(CODER_SYSTEM_PROMPT)
-    }
-
-    fn assemble_rendered_context(&mut self, scope_instruction: &str) -> String {
-        let assembler = ContextAssembler::default();
-        let render_assembler = assembler.clone();
-        let operator = ToolOperator::new(self.working_dir.clone());
-        let scope_instruction_for_task = scope_instruction.to_string();
-        let assembled = block_on_context_task(async move {
-            tokio::task::spawn_blocking(move || {
-                assembler.assemble(&scope_instruction_for_task, &operator)
-            })
-            .await
-            .map_err(|error| anyhow::anyhow!("failed to join context assembly task: {error}"))?
-        })
-        .ok();
-        if let Some(context) = assembled.clone() {
-            self.last_assembled_context = Some(context);
-        }
-        assembled
-            .as_ref()
-            .map(|context| render_assembler.render(context))
-            .unwrap_or_else(|| "## Context\n[context: unavailable]\n".to_string())
-    }
-
-    fn resolved_notes_path(&self) -> Option<PathBuf> {
-        resolve_notes_path_for_write(self.notes_path.as_deref())
-    }
-
-    fn resolved_existing_notes_path(&self) -> Option<PathBuf> {
-        resolve_notes_path_for_read(self.notes_path.as_deref())
     }
 }
